Move file to Root folder
